feat(authenticate-service): add filter to modify payload
I needed to extend the returned payload from the plugin. The plugin only allows you to choose from a fixed number of payload options. By adding in a filter called `simple_jwt_login_generate_payload` we can add in new payload values which get added to the generated JSON Web Token. I needed this to add a user avatar to the payload.